requirements.txt
pandas
requests>=2.31.0
apache-airflow
python-dotenv>=1.0.0
pytest==7.2.1
flake8==5.0.4
kafka-python==2.0.2
confluent-kafka==2.1.1
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7
